requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.2.1
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
distro==1.9.0
docutils==0.21.2
filetype==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jieba==0.42.1
jiter==0.10.0
Kivy==2.3.1
Kivy-Garden==0.1.5
openai==1.84.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
requests==2.32.4
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
# HTML 提取依赖（可选，按优先级排序）
trafilatura>=1.6.0  # 推荐：最好的 HTML 正文提取库
readability-lxml>=0.8.1  # Fallback 1：基于 readability 的提取
beautifulsoup4>=4.12.0  # Fallback 2：手动 HTML 解析
lxml>=4.9.0  # BeautifulSoup 的 XML/HTML 解析器
